+++ src/main/java/org/folio/circulation/domain/PatronGroupRepository.java
@@ -1,5 +1,8 @@
 package org.folio.circulation.domain;
 
+import static java.util.concurrent.CompletableFuture.completedFuture;
+import static org.folio.circulation.support.HttpResult.succeeded;
+
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
@@ -47,6 +50,10 @@ class PatronGroupRepository {
       .distinct()
       .collect(Collectors.toList());
 
+    if (groupsToFetch.isEmpty()) {
+      return completedFuture(succeeded(multipleRequests));
+    }
+
     final String query = CqlHelper.multipleRecordsCqlQuery(groupsToFetch);
 
     return patronGroupsStorageClient.getMany(query, groupsToFetch.size(), 0)
@@ -54,7 +61,7 @@ class PatronGroupRepository {
       .thenApply(multiplePatronGroupsResult -> multiplePatronGroupsResult.next(
         patronGroups -> matchGroupsToUsers(multipleRequests, patronGroups)));
   }
-    
+
   private HttpResult<MultipleRecords<PatronGroup>> mapResponseToPatronGroups(Response response) {
     return MultipleRecords.from(response, PatronGroup::from, "usergroups");
   }