+++ src/main/java/org/folio/circulation/domain/Request.java
@@ -29,6 +29,8 @@ public class Request implements ItemRelatedRecord, UserRelatedRecord {
   private final Loan loan;
   private final ServicePoint pickupServicePoint;
 
+  public static final String REQUEST_TYPE = "requestType";
+
   private boolean changedPosition = false;
 
   public Request(
@@ -124,7 +126,7 @@ public class Request implements ItemRelatedRecord, UserRelatedRecord {
         pickupServicePoint);
   }
 
-  Request withLoan(Loan newLoan) {
+  public Request withLoan(Loan newLoan) {
     return new Request(representation, item, requester, proxy, newLoan,
         pickupServicePoint);
   }
@@ -156,7 +158,7 @@ public class Request implements ItemRelatedRecord, UserRelatedRecord {
   }
 
   RequestType getRequestType() {
-    return RequestType.from(representation.getString("requestType"));
+    return RequestType.from(representation.getString(REQUEST_TYPE));
   }
 
   Boolean allowedForItem() {