+++ src/main/java/org/folio/circulation/domain/UserRepository.java
@@ -1,5 +1,6 @@
 package org.folio.circulation.domain;
 
+import static java.util.concurrent.CompletableFuture.completedFuture;
 import static org.folio.circulation.support.HttpResult.failed;
 import static org.folio.circulation.support.HttpResult.succeeded;
 import static org.folio.circulation.support.ValidationErrorFailure.failure;
@@ -57,7 +58,7 @@ public class UserRepository {
   public CompletableFuture<HttpResult<User>> getProxyUserByBarcode(String barcode) {
     //Not proxying, so no need to get proxy user
     if(StringUtils.isBlank(barcode)) {
-      return CompletableFuture.completedFuture(succeeded(null));
+      return completedFuture(succeeded(null));
     }
     else {
       return getUserByBarcode(barcode, "proxyUserBarcode");
@@ -79,7 +80,7 @@ public class UserRepository {
         .next(user -> user.map(HttpResult::succeeded).orElseGet(() -> failed(failure(
           "Could not find user with matching barcode", propertyName, barcode)))));
   }
-  
+
   CompletableFuture<HttpResult<MultipleRecords<Request>>> findUsersForRequests(
     MultipleRecords<Request> multipleRequests) {
 
@@ -90,6 +91,10 @@ public class UserRepository {
       .flatMap(Collection::stream)
       .distinct()
       .collect(Collectors.toList());
+
+    if (usersToFetch.isEmpty()) {
+      return completedFuture(succeeded(multipleRequests));
+    }
 
     final String query = CqlHelper.multipleRecordsCqlQuery(usersToFetch);