+++ src/test/java/api/requests/RequestsAPICreationTests.java
@@ -256,13 +256,19 @@ public class RequestsAPICreationTests extends APITests {
     MalformedURLException {
 
     UUID itemId = UUID.randomUUID();
+    UUID patronId = usersFixture.charlotte().getId();
+    final UUID pickupServicePointId = servicePointsFixture.cd1().getId();
 
+    //Check RECALL -- should give the same response when placing other types of request.
     Response postResponse = requestsClient.attemptCreate(new RequestBuilder()
       .recall()
       .withItemId(itemId)
-      .withRequesterId(usersFixture.charlotte().getId()));
+      .withPickupServicePointId(pickupServicePointId)
+      .withRequesterId(patronId));
 
     assertThat(postResponse, hasStatus(HTTP_VALIDATION_ERROR));
+    assertThat(postResponse.getJson(), hasErrorWith(allOf(
+      hasMessage("Item does not exist"))));
   }
 
   @Test
@@ -527,7 +533,7 @@ public class RequestsAPICreationTests extends APITests {
   }
 
   @Test
-  public void creatingARequestDoesNotStoreRequesterInformationWhenUserNotFound()
+  public void cannotCreateRequestWithUserBelongingToNoPatronGroup()
     throws InterruptedException,
     ExecutionException,
     TimeoutException,
@@ -536,28 +542,27 @@ public class RequestsAPICreationTests extends APITests {
     final IndividualResource smallAngryPlanet = itemsFixture.basedUponSmallAngryPlanet();
     final IndividualResource jessica = usersFixture.jessica();
     final UUID pickupServicePointId = servicePointsFixture.cd1().getId();
+    final IndividualResource noUserGroupBob = usersFixture.noUserGroupBob();
 
     loansFixture.checkOut(smallAngryPlanet, jessica);
 
-    UUID nonExistentRequester = UUID.randomUUID();
-
     DateTime requestDate = new DateTime(2017, 7, 22, 10, 22, 54, DateTimeZone.UTC);
 
-    IndividualResource createdRequest = requestsFixture.place(new RequestBuilder()
+    final Response recallResponse = requestsClient.attemptCreate(new RequestBuilder()
       .recall()
-      .withRequestDate(requestDate)
       .forItem(smallAngryPlanet)
       .withPickupServicePointId(pickupServicePointId)
-      .withRequesterId(nonExistentRequester));
+      .withRequestDate(requestDate)
+      .by(noUserGroupBob));
 
-    JsonObject representation = createdRequest.getJson();
+    assertThat(recallResponse, hasStatus(HTTP_VALIDATION_ERROR));
 
-    assertThat("has no information for missing requesting user",
-      representation.containsKey("requester"), is(false));
+    assertThat(recallResponse.getJson(), hasErrorWith(allOf(
+      hasMessage("A valid patron group is required. PatronGroup ID is null"))));
   }
 
   @Test
-  public void creatingARequestStoresItemInformationWhenRequestingUserNotFound()
+  public void cannotCreateRequestWithoutValidUser()
     throws InterruptedException,
     ExecutionException,
     TimeoutException,
@@ -573,28 +578,17 @@ public class RequestsAPICreationTests extends APITests {
 
     DateTime requestDate = new DateTime(2017, 7, 22, 10, 22, 54, DateTimeZone.UTC);
 
-    IndividualResource createdRequest = requestsFixture.place(new RequestBuilder()
+    final Response recallResponse = requestsClient.attemptCreate(new RequestBuilder()
       .recall()
-      .withRequestDate(requestDate)
       .forItem(smallAngryPlanet)
       .withPickupServicePointId(pickupServicePointId)
+      .withRequestDate(requestDate)
       .withRequesterId(nonExistentRequesterId));
 
-    JsonObject representation = createdRequest.getJson();
+    assertThat(recallResponse, hasStatus(HTTP_VALIDATION_ERROR));
 
-    assertThat("has information taken from item",
-      representation.containsKey("item"), is(true));
-
-    assertThat("title is taken from item",
-      representation.getJsonObject("item").getString("title"),
-      is("The Long Way to a Small, Angry Planet"));
-
-    assertThat("barcode is taken from item",
-      representation.getJsonObject("item").getString("barcode"),
-      is("036000291452"));
-
-    assertThat("has no information for missing requesting user",
-      representation.containsKey("requester"), is(false));
+    assertThat(recallResponse.getJson(), hasErrorWith(allOf(
+      hasMessage("A valid user and patron group are required. User is null"))));
   }
 
   @Test
@@ -910,7 +904,7 @@ public class RequestsAPICreationTests extends APITests {
 
     assertThat(pagedRequest, hasStatus(HTTP_VALIDATION_ERROR));
     JsonArray errors = pagedRequest.getJson().getJsonArray("errors");
-    assertThat(errors.getJsonObject(0).getString("message").toLowerCase(), is("item is " + ItemStatus.CHECKED_OUT.toString().toLowerCase()));
+    assertThat(errors.getJsonObject(0).getString("message").toLowerCase(), is (("Page requests are not allowed for " + ItemStatus.CHECKED_OUT.getValue() + " item status combination").toLowerCase()));
   }
 
   @Test
@@ -934,7 +928,7 @@ public class RequestsAPICreationTests extends APITests {
 
     assertThat(pagedRequest2, hasStatus(HTTP_VALIDATION_ERROR));
     JsonArray errors = pagedRequest2.getJson().getJsonArray("errors");
-    assertThat(errors.getJsonObject(0).getString("message").toLowerCase(), is("item is " + ItemStatus.AWAITING_PICKUP.toString().toLowerCase()));
+    assertThat(errors.getJsonObject(0).getString("message").toLowerCase(), is (("Page requests are not allowed for " + ItemStatus.AWAITING_PICKUP.getValue() + " item status combination").toLowerCase()));
   }
 
   @Test
@@ -957,7 +951,7 @@ public class RequestsAPICreationTests extends APITests {
 
     assertThat(pagedRequest2, hasStatus(HTTP_VALIDATION_ERROR));
     JsonArray errors = pagedRequest2.getJson().getJsonArray("errors");
-    assertThat(errors.getJsonObject(0).getString("message").toLowerCase(), is("item is " + ItemStatus.PAGED.toString().toLowerCase()));
+    assertThat(errors.getJsonObject(0).getString("message").toLowerCase(), is(("Page requests are not allowed for " + ItemStatus.PAGED.getValue() + " item status combination").toLowerCase()));
   }
 
   @Test
@@ -982,7 +976,7 @@ public class RequestsAPICreationTests extends APITests {
 
     assertThat(pagedRequest2, hasStatus(HTTP_VALIDATION_ERROR));
     JsonArray errors = pagedRequest2.getJson().getJsonArray("errors");
-    assertThat(errors.getJsonObject(0).getString("message").toLowerCase(), is("item is " + ItemStatus.IN_TRANSIT.toString().toLowerCase()));
+    assertThat(errors.getJsonObject(0).getString("message").toLowerCase(), is (("Page requests are not allowed for " + ItemStatus.IN_TRANSIT.getValue() + " item status combination").toLowerCase()));
   }
 
   @Test
@@ -1077,7 +1071,7 @@ public class RequestsAPICreationTests extends APITests {
 
     assertThat(recallResponse, hasStatus(HTTP_VALIDATION_ERROR));
     JsonArray errors = recallResponse.getJson().getJsonArray("errors");
-    assertThat(errors.getJsonObject(0).getString("message").toLowerCase(), is("item is " + ItemStatus.AVAILABLE.toString().toLowerCase()));
+    assertThat(errors.getJsonObject(0).getString("message").toLowerCase(), is (("Recall requests are not allowed for " + ItemStatus.AVAILABLE.getValue() + " item status combination").toLowerCase()));
   }
 
   @Test
@@ -1098,7 +1092,7 @@ public class RequestsAPICreationTests extends APITests {
 
     assertThat(holdRequest, hasStatus(HTTP_VALIDATION_ERROR));
     JsonArray errors = holdRequest.getJson().getJsonArray("errors");
-    assertThat(errors.getJsonObject(0).getString("message").toLowerCase(), is("item is " + ItemStatus.MISSING.toString().toLowerCase()));
+    assertThat(errors.getJsonObject(0).getString("message").toLowerCase(), is (("Recall requests are not allowed for " + ItemStatus.MISSING.getValue() + " item status combination").toLowerCase()));
   }
 
   @Test
@@ -1119,7 +1113,7 @@ public class RequestsAPICreationTests extends APITests {
 
     assertThat(recallResponse, hasStatus(HTTP_VALIDATION_ERROR));
     JsonArray errors = recallResponse.getJson().getJsonArray("errors");
-    assertThat(errors.getJsonObject(0).getString("message").toLowerCase(), is("item is " + ItemStatus.PAGED.toString().toLowerCase()));
+    assertThat(errors.getJsonObject(0).getString("message").toLowerCase(), is (("Recall requests are not allowed for " + ItemStatus.PAGED.getValue() + " item status combination").toLowerCase()));
   }
 
   @Test
@@ -1259,8 +1253,9 @@ public class RequestsAPICreationTests extends APITests {
 
     assertThat(recallResponse, hasStatus(HTTP_VALIDATION_ERROR));
     JsonArray errors = recallResponse.getJson().getJsonArray("errors");
-    assertThat(errors.getJsonObject(0).getString("message").toLowerCase(), is("item is " + ItemStatus.AVAILABLE.toString().toLowerCase()));
+    assertThat(errors.getJsonObject(0).getString("message").toLowerCase(), is(("Hold requests are not allowed for " + ItemStatus.AVAILABLE.getValue() + " item status combination").toLowerCase()));
   }
+
 
   public static IndividualResource setupPagedItem(IndividualResource requestPickupServicePoint, ItemsFixture itemsFixture,
                                                   ResourceClient requestClient, UsersFixture usersFixture)